Add spin mutexes support
Add spin_mutex and shared_spin_mutex as more efficient replacements for
std::mutex and std::shared_mutex (when the scopes protected are very
small). They both spin with the help of spin_backoff.

Add spin_backoff abstraction to be able to spin with an exponential
backoff. The spin is made, as much as possible with "pause" instructions
to put the CPU in the low-energy state. We attempt to exponentially
increase the amount of spinning we are doing each round. After a certain
threshold we cease the CPU quanta to another thread.