Workers use a special task list
Instead of using a queue for the worker's task list, we use a special
stack. Pushing and popping is done at the front of the stack. But, this
task also allows stealing tasks from the bottom of the queue.

This way, the worker will work on the most "local" task possible: the
last one added. On the other hand, if somebody will steal a task it will
steal the task with the lowest locality.