Spawn wakes up only one worker thread
Instead of waking up all the threads, just in case, we only wake up one
worker thread. First, we check if there is one preparing to go to idle,
and if there is no such think try to wake one (and only one) worker that
is idle.

This improves some performance (mainly for latency).