Add support for spawn_and_wait
We create a task_contorl object that will constantly be checked for
completion. If we have a worker data, we try to execute tasks from it,
hoping to bring that task to completion asap. Otherwise, just wait.

When calling spawn_and_wait from a worker thread, the regular spawning
and execution patten happens.

When calling spawn_and_wait from a non-worker thread, we temporary try
to create a worker_thread_data structure for it, and make this thread
temporary join our task system. As soon as the task(s) is/are completed
the new thread leaves our task system.

For performance reasons, we have a limited number of slots reserved for
external threads.