Serializers can have 2 executors
A serializer can be constructed with 2 executors: one for enqueueing new
tasks, and one for continuing executing tasks after one task is
finished.

If the first executor is not given, the global_executor will be used.
If the second executor is not given, but the first one is, the first
executor will be used to enqueue continuations.
If no executor is given, a spwan_continuation_executor will be used to
enqueue continuations.

Move serializer implementation in .cpp files.

Small refactoring for executing tasks.