Fix race condition in conc_scan
The joins for the last line were not waited on by the main algorithm.
Actually, at a closer inspection, we don't need the last line operations
at all. So, removed any tasks from the last line.

Also made it possible to manually clear the dependencies of the tasks
in a task graph. Made a test not leak memory.