Allow pipeline object to be used in the stages
So far, the pipeline object would only be available after construction,
after all the stages would be created. That would make it hard for the
stages to use the pipeline objects, thus making it hard for various
stages to push items in the pipeline.

Also added test to ensure that the line item can be received by
reference, and modified between stages.