Fix "finish_task" finishing before the term
When using get_continuation() with task spawning, there might be a race
condition, as the task can be executed and continuation called before
the main thread reaches to the next instruction.

To fix this with finish_wait we use an initial increment of 1, so that
when the wait is actually called we know we should have >= 1 ref count.

For the other case, explicitly document the exception.

Also allow get_continuation() to take a count parameter, so that the
continuation object can be reused.