Task executors can be noexcept
Define a concept task_executor that refines the executor concept and
also adds the following constraint:
- can execute object of type concore::task
- the execution of these objects is noexcept

The expectation is that whenever there is an exception thrown while
trying to enqueue a task (i.e., before the task is started to be
executed), the exception is caught and the continuation handler in the
task is called with the appropriate exception.

Make sure that the global executor and the spawn executors model the
task_executor concept.

Add test to check this.